+++ _docs/_developer_guide/platform_integration_guides/android/changelog.md
@@ -1,12 +1,15 @@
 ---
 nav_title: Changelog
+article_title: Changelog for Android and FireOS
 page_order: 30
 page_type: update
-platform: android
-description: "This page lists updates to Braze's Android SDK changelog."
+platform: 
+  - Android
+  - FireOS
+description: "This page lists updates to Braze's Android/FireOS SDK changelog."
 
 ---
 
-# Android SDK Changelog
+# Android SDK changelog
 
 {% markdown_embed https://raw.githubusercontent.com/Appboy/appboy-android-sdk/master/CHANGELOG.md %}